requirements.txt
fastapi
uvicorn[standard]
httpx
selectolax
sqlmodel
duckdb  # or sqlite
pydantic>=2.0
python-dotenv
pytest
httpx[http2]
ruff
pre-commit
fastapi
uvicorn
httpx
pandas
plotly
fastapi-cache2